+++ src/controllers/shelter.controllers.js
@@ -1,11 +1,15 @@
-import shelters from '../models/Shelter';
+import shelters from '../models/shelter.model';
 
 export const getShelters = async (req, res) => {
   try {
     const sheltersList = await shelters.find(
       {},
       {
-        place_name: 1, address: 1, contact_person: 1, phone: 1, locs: 1,
+        place_name: 1,
+        address: 1,
+        contact_person: 1,
+        phone: 1,
+        locs: 1,
       },
     );
     res.status(200).send(sheltersList);
