+++ src/routes/auth.js
@@ -5,7 +5,7 @@ const jwt = require('jsonwebtoken');
 
 const router = express.Router();
 
-const User = require('../models/User');
+const User = require('../models/user.model');
 
 const validate = [
   check('username')
@@ -17,10 +17,11 @@ const validate = [
     .withMessage('Your password must be at least eight charachters'),
 ];
 
-const generateToken = (user) => jwt.sign(
-  { _id: user._id, email: user.email, username: user.username },
-  'SECRET_KEY',
-);
+const generateToken = (user) =>
+  jwt.sign(
+    { _id: user._id, email: user.email, username: user.username },
+    'SECRET_KEY',
+  );
 
 const loginValidation = [
   check('email').isEmail().withMessage('Please provide a valid email address'),
